Add Jest testing setup
