+++ client/src/pages/__tests__/Alerts.test.tsx
@@ -1,0 +1,7 @@
+import { render, screen } from '@testing-library/react';
+import Alerts from '../Alerts';
+
+test('renders Alerts heading', () => {
+  render(<Alerts />);
+  expect(screen.getByText('Alerts')).toBeInTheDocument();
+});
