+++ server/app.js
@@ -1,0 +1,26 @@
+const express = require('express');
+const cors = require('cors');
+const fetch = (...args) =>
+  import('node-fetch').then(({ default: fetch }) => fetch(...args));
+
+const app = express();
+app.use(cors());
+app.use(express.json());
+
+app.get('/', (req, res) => {
+  res.json({ status: 'ok' });
+});
+
+app.get('/api/prices', async (req, res) => {
+  try {
+    const { token } = req.query;
+    const response = await fetch(`https://api.dexscreener.com/latest/dex/tokens/${token}`);
+    const data = await response.json();
+    res.json(data);
+  } catch (err) {
+    console.error(err);
+    res.status(500).json({ error: 'Failed to fetch price data' });
+  }
+});
+
+module.exports = app;
